+++ src/control_plane/state/mod.rs
@@ -1,9 +1,17 @@
 mod project_record;
+mod resource_lifecycle;
+mod resource_record;
+mod resource_record_options;
+mod resource_retention;
 mod sqlite_state_store;
 mod state_store;
 mod state_store_error;
 
 pub(crate) use project_record::ProjectRecord;
+pub(crate) use resource_lifecycle::ResourceLifecycle;
+pub(crate) use resource_record::ResourceRecord;
+pub(crate) use resource_record_options::ResourceRecordOptions;
+pub(crate) use resource_retention::ResourceRetention;
 pub(crate) use sqlite_state_store::SqliteStateStore;
 pub(crate) use state_store::StateStore;
 pub(crate) use state_store_error::StateStoreError;